Time: 35 ms (77.9%), Space: 105.8 MB (63.56%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int maxFrequency(vector<int>& nums, int k) {
         sort(nums.begin(),nums.end());
-        int left=0; long long sum=0;
-        int ans=1;
-        for(int right=0;right<nums.size();right++){
+        long long left=0; long long sum=0;
+        long long ans=1;
+        for(long long right=0;right<nums.size();right++){
 sum+=nums[right];
 while((long long)nums[right]*(right-left+1)-sum>k){
     sum-=nums[left];
